# public/swagger_v4.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://api.bible.build/v2",
      "description": "Online Staging server for testing"
    },
    {
      "url": "https://api.dbp.dev/v2",
      "description": "Internal Staging server for testing"
    }
  ],
  "info": {
    "title": "Digital Bible Platform [Koinos]",
    "description": "This is a simple API",
    "version": "4.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://dbs.org/contact",
      "email": "jon@dbs.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/alphabets": {
      "get": {
        "operationId": "v4_alphabets.all",
        "summary": "Get Alphabets",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Wiki"
        ],
        "responses": {
          "200": {
            "description": "An array of Alphabets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alphabets"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "v4_alphabets.store",
        "summary": "Create a new Alphabet",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Wiki"
        ],
        "responses": {
          "200": {
            "description": "An Success Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Alphabet Successfully created"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/alphabets/{id}": {
      "get": {
        "operationId": "v4_alphabets.one",
        "summary": "A single Alphabet",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Wiki"
        ],
        "responses": {
          "200": {
            "description": "An array of Alphabets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alphabets"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alphabets": {
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Adlam Script"
          },
          "script": {
            "type": "string",
            "example": "Adlm"
          },
          "family": {
            "type": "string",
            "example": "African"
          },
          "type": {
            "type": "string",
            "example": "alphabet"
          },
          "direction": {
            "type": "string",
            "example": "rtl"
          }
        }
      },
...
    },
    "parameters": {
      "v": {
        "name": "v",
        "in": "query",
        "description": "The Version Number of the API",
        "required": true,
        "example": 4,
        "schema": {
          "type": "string"
        }
      },
      "key": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "pretty": {
        "name": "pretty",
        "in": "query",
        "description": "A sugar param, simply will format the call for human eyes",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "version": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "reply": {
        "name": "reply",
        "in": "query",
        "description": "Specifies the response type requested by the caller. Default is json.",
        "schema": {
          "type": "string",
          "enum": [
            "json",
            "html",
            "jsonp",
            "yaml",
            "xml"
          ]
        }
      },
...
    }
  }
}